add to Job Position Api delete and update services

+++ src/main/java/com/umutdoruk/hrms/repository/JobPositionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface JobPositionRepository extends JpaRepository<JobPosition, Integer> {
+public interface JobPositionRepository extends JpaRepository<JobPosition, Long> {
     List<JobPosition> findByName(String jobPositionName);
 }
